packages/api/pyproject.toml
[project]
name = "lumen-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.3",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "supabase>=2.15.0",
    "uvicorn[standard]>=0.42.0",
    "python-multipart>=0.0.20",
    # ML Pipeline dependencies
    "opencv-python-headless>=4.8.0",
    "ultralytics>=8.1.0",
    "supervision>=0.20.0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "lightgbm>=4.0.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
]
